refactor(parser): reformat code to improve readability and maintainability feat(parser): add support for creating tables with createTablStmt feat(ColumnVariable, RowVariable): add set_value method to allow for changing the value of a variable The code has been reformatted to improve readability and maintainability. The createTablStmt function has been added to support creating tables. The set_value method has been added to the ColumnVariable and RowVariable classes to allow for changing the value of a variable.
feat(TableVariable.py): add support for creating tables with a specified number of rows and columns
feat(llvm_compile.py): add support for printing tables
The TableVariable class now has a constructor that takes in the number of rows and columns as arguments, allowing for the creation of tables with a specified size. The create_table method in the ProgramCompiler class now takes in the number of rows and columns as arguments and creates a table with the specified size. The llvm_compile.py file now has support for printing tables using the newly added print_table function.